Added exceptions and errors handling
