+++ conf/init.properties.example
@@ -2,5 +2,6 @@ DRILL_HOST=http://localhost:8047
 DRIVER_TYPE=CHROME_LINUX
 WEBDRIVERS_PATH=webdrivers/chromedriver_linux64
 SECURE_DRILL=true
+DEFAULT_TIMEOUT=2
 DRILL_USER_NAME=mapr
 DRILL_USER_PASSWORD=mapr
