+++ conf/init.properties_Windows_Chrome
@@ -2,5 +2,6 @@ DRILL_HOST=http://localhost:8047
 DRIVER_TYPE=CHROME
 WEBDRIVERS_PATH=webdrivers/chromedriver_win32.exe
 SECURE_DRILL=true
+DEFAULT_TIMEOUT=2
 DRILL_USER_NAME=mapr
 DRILL_USER_PASSWORD=mapr
