+++ conf/init.properties_Windows_EDGE
@@ -2,5 +2,6 @@ DRILL_HOST=http://localhost:8047
 DRIVER_TYPE=EDGE
 WEBDRIVERS_PATH=webdrivers/MicrosoftWebDriver_EDGE.exe
 SECURE_DRILL=true
+DEFAULT_TIMEOUT=2
 DRILL_USER_NAME=mapr
 DRILL_USER_PASSWORD=mapr
